[NCL-5267] Log PNC build related stages and duration
The build stages logged are:

- Scheduling (when a build config is being broken down into build tasks)
- Enqueued (the amount of time the build stayed in enqueued state)

- The various stages in `BuildExecutionStatus`
  * NEW,
  * REPO_SETTING_UP,
  * BUILD_ENV_SETTING_UP,
  * BUILD_ENV_WAITING,
  * BUILD_ENV_SETUP_COMPLETE_SUCCESS,
  * BUILD_SETTING_UP,
  * BUILD_WAITING,
  * COLLECTING_RESULTS_FROM_BUILD_DRIVER,
  * COLLECTING_RESULTS_FROM_REPOSITORY_MANAGER,
  * REPOSITORY_MANAGER_SEALING_TRACKING_RECORD,
  * REPOSITORY_MANAGER_DOWNLOADING_TRACKING_REPORT,
  * REPOSITORY_MANAGER_PROCESSING_DEPENDENCIES,
  * REPOSITORY_MANAGER_PROCESSING_BUILT_ARTIFACTS,
  * REPOSITORY_MANAGER_REMOVE_BUILD_AGGREGATION_GROUP,
  * REPOSITORY_MANAGER_PROMOTION_BUILD_CONTENT_SET,
  * COLLECTING_RESULTS_FROM_REPOSITORY_MANAGER_COMPLETED_SUCCESS,
  * BUILD_COMPLETED_SUCCESS,
  * BUILD_ENV_DESTROYING,
  * BUILD_ENV_DESTROYED,
  * FINALIZING_EXECUTION,

+++ common/src/main/java/org/jboss/pnc/common/util/BuildStageUtils.java
@@ -37,7 +37,7 @@ public class BuildStageUtils {
      * @param buildStage build stage who's duration we're logging
      * @param duration in seconds
      */
-    public static void logBuildStage(String buildStage, int duration) {
+    public static void logBuildStage(String buildStage, long duration) {
 
         Map<String, String> mdc = MDC.getCopyOfContextMap();
         if (mdc == null) {
